Issue #108: Mark ApiGateway certificates as CA certs
AWS API Gateway certficates are self-signed but don't have
the IsCA flag set to true and also don't mark the certificate
as to be used for certificate signing. This prevents Go to
accept the certificate for client-cert authentication. Since
it isn't possible to use a custom certificate for client authentication
in the AWS API Gateway we need to patch the certificate on the fly.

This is a preliminary patch to verify that the approach works but
has the the certificate detection hard-coded and permanent. This
should be configurable and off by default.