Issue #1: Add TCP proxy with SNI support
This patch adds support for configuring a listener
to provide a passthrough TCP proxy for TLS connections
and use the SNI server extension in the ClientHello
message for routing. Targets need to register their
prefix as 'host/' for now.

This implementation is functional but not production
ready since it does not yet set proper timeouts on
connections. This will be fixed before the change
is merged.